Updating checking logic

+++ src/services/ProcessorService.js
@@ -76,17 +76,19 @@ async function associateChallengeTerms (v5Terms, legacyChallengeId) {
   logger.debug(`Legacy Terms ${JSON.stringify(legacyTermsArray)}`)
   logger.debug(`Legacy NDA Found ${JSON.stringify(legacyNDA)}`)
 
-  if (nda.id && !legacyNDA) {
+  if (nda && nda.id && !legacyNDA) {
     logger.debug('v5 NDA exist, not in legacy. Adding to Legacy.')
     const m2mToken = await helper.getM2MToken()
     const v5Term = await getV5Terms(nda.id, m2mToken)
     return termsService.addTermsToChallenge(legacyChallengeId, v5Term.legacyId, config.LEGACY_SUBMITTER_ROLE_ID)
   }
 
-  if (!nda && legacyNDA) {
+  if (!nda && legacyNDA && legacyNDA.id) {
     logger.debug('Legacy NDA exist, not in V5. Removing from Legacy.')
     return termsService.removeTermsFromChallenge(legacyChallengeId, legacyNDA.id, config.LEGACY_SUBMITTER_ROLE_ID)
   }
+
+  logger.debug('Associate Challenge Terms - Nothing to Do')
 }
 
 /**
